backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
PyJWT==2.10.1
bcrypt==4.2.1
mysql-connector-python==9.2.0
pytesseract==0.3.13
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
werkzeug==3.1.3
